+++ Makefile
@@ -79,7 +79,7 @@ build:
 
 build-cli:
 	@echo "Building WhisperKit CLI..."
-	@swift build -c release --product whisper-kit
+	@swift build -c release --product transcribe
 
 
 test:
